Allow unidling to ignore configurable user agent strings
When undiling, it may be necessary to ignore certain user agent strings to avoid unidling gears that do not need to be unidled. For example, a request from a search engine spider should not cause the application to unidle. This change adds the ability to configure a list of strings that will be checked against the user agent string of each request and will cause the unilder to ignore the request if there is a match.

Bug 1305166
https://bugzilla.redhat.com/show_bug.cgi?id=1305166